implementacao de tratamento de erros

+++ produto-crud/src/main/java/com/desafio/dio/produto_crud/service/ProdutoService.java
@@ -1,5 +1,6 @@
 package com.desafio.dio.produto_crud.service;
 
+import com.desafio.dio.produto_crud.exception.ProdutoNotFoundException;
 import com.desafio.dio.produto_crud.model.Produto;
 import com.desafio.dio.produto_crud.repository.ProdutoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,7 +20,7 @@ public class ProdutoService {
     }
 
     public Optional<Produto> findById(Long id) {
-        return produtoRepository.findById(id);
+        return Optional.ofNullable(produtoRepository.findById(id).orElseThrow(() -> new ProdutoNotFoundException(id)));
     }
 
     public Produto save(Produto produto) {
@@ -27,6 +28,9 @@ public class ProdutoService {
     }
 
     public void deleteById(Long id) {
+        if (!produtoRepository.existsById(id)) {
+            throw new ProdutoNotFoundException(id);
+        }
         produtoRepository.deleteById(id);
     }
 }
